댓글 작성 화면, test

+++ springedu/src/main/java/com/kh/portfolio/board/dao/RboardDAO.java
@@ -20,7 +20,7 @@ public interface RboardDAO {
 	//대댓글 작성
 	int reply(RboardVO rboardVO);	
 	//이전step update
-	int updateStep(int rgroup, int rstep);
+	int updateStep(long rgroup, long rstep);
 	//댓글 호감, 비호감	
 	//투표이력 없으면 추가 있으면 변경
 	int vote(VoteVO voteVO);
